+++ engine/language_client_python/python_src/baml_py/ctx_manager.py
@@ -40,31 +40,27 @@ class CtxManager:
             ctx[thread_id] = self.rt.create_context_manager()
         return ctx[thread_id]
 
-    
     def allow_reset(self) -> bool:
         ctx = self.ctx.get()
-        
+
         if len(ctx) > 1:
             print("Too many ctxs!")
             return False
 
-
         thread_id = current_thread_id()
         if thread_id not in ctx:
             print("Thread not in ctx!")
-            return False        
+            return False
 
         for c in ctx.values():
             if c.context_depth() > 0:
                 print("Context depth is greater than 0!")
                 return False
-    
 
         return True
 
     def reset(self) -> None:
         self.ctx.set({current_thread_id(): self.rt.create_context_manager()})
-        
 
     def upsert_tags(self, **tags: str) -> None:
         mngr = self.__ctx()
