+++ engine/language_client_python/src/errors.rs
@@ -17,15 +17,15 @@ create_exception!(baml_py, BamlClientHttpError, BamlClientError);
 
 // Define the BamlValidationError exception with additional fields
 // can't use extends=PyException yet https://github.com/PyO3/pyo3/discussions/3838
-#[pyfunction]
+
+#[allow(non_snake_case)]
 fn raise_baml_validation_error(prompt: String, message: String, raw_output: String) -> PyErr {
     Python::with_gil(|py| {
-        // Import the current module to access the BamlValidationError class
-        let module = PyModule::import(py, "baml_py.errors").unwrap();
-        let exception = module.getattr("BamlValidationError").unwrap();
+        let internal_monkeypatch = py.import("baml_py.internal_monkeypatch").unwrap();
+        let exception = internal_monkeypatch.getattr("BamlValidationError").unwrap();
         let args = (prompt, message, raw_output);
-        let instance = exception.call1(args).unwrap();
-        PyErr::from_value(instance.into())
+        let inst = exception.call1(args).unwrap();
+        PyErr::from_value(inst)
     })
 }
 
@@ -33,60 +33,6 @@ fn raise_baml_validation_error(prompt: String, message: String, raw_output: Stri
 /// IIRC the name of this function is the name of the module that pyo3 generates (errors.py)
 #[pymodule]
 pub fn errors(parent_module: &Bound<'_, PyModule>) -> PyResult<()> {
-    // Define the BamlValidationError Python exception class in a hacky way first, manually into that errors module.
-    let errors_module = PyModule::from_code_bound(
-        parent_module.py(),
-        r#"
-class BamlValidationError(Exception):
-    def __init__(self, prompt, message, raw_output):
-        super().__init__(message)
-        self.prompt = prompt
-        self.message = message
-        self.raw_output = raw_output
-
-    def __str__(self):
-        return f"BamlValidationError(message={self.message}, raw_output={self.raw_output}, prompt={self.prompt})"
-
-    def __repr__(self):
-        return f"BamlValidationError(message={self.message}, raw_output={self.raw_output}, prompt={self.prompt})"
-"#,
-        "errors.py",
-        "errors",
-    )?;
-
-    // Add the raise_baml_validation_error function to the module
-    parent_module.add_wrapped(wrap_pyfunction!(raise_baml_validation_error))?;
-
-    // add the other exceptions in
-    errors_module.add(
-        "BamlError",
-        errors_module.py().get_type_bound::<BamlError>(),
-    )?;
-    errors_module.add(
-        "BamlInvalidArgumentError",
-        errors_module
-            .py()
-            .get_type_bound::<BamlInvalidArgumentError>(),
-    )?;
-    errors_module.add(
-        "BamlClientError",
-        errors_module.py().get_type_bound::<BamlClientError>(),
-    )?;
-    errors_module.add(
-        "BamlClientHttpError",
-        errors_module.py().get_type_bound::<BamlClientHttpError>(),
-    )?;
-
-    parent_module.add_submodule(&errors_module)?;
-
-    // we have to do this hack or python will complain the baml_py.errors is not a package.
-    parent_module
-        .py()
-        .import("sys")?
-        .getattr("modules")?
-        .set_item("baml_py.errors", errors_module.clone())?;
-
-    // now add the other errors again to the parent module
     parent_module.add(
         "BamlError",
         parent_module.py().get_type_bound::<BamlError>(),
